pyproject.toml
[project]
authors = [
    { name = "Tomoyuki Suzuki", email = "suzuki_tomoyuki@cyberagent.co.jp" },
]
name = "layerd"
version = "0.1.0"
description = "LayerD: Decomposing Raster Graphic Designs into Layers"
requires-python = ">=3.10.0"
dependencies = [
    "numpy>=1.26",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "opencv-python>=4.5.3",
    "scikit-image>=0.25.0",
    "tqdm>=4.62.0",
    "huggingface-hub>=0.25.0",
    "accelerate>=1.1.1",
    "cr-renderer",
    "hydra-core>=1.3.2",
    "simple-lama-inpainting>=0.1.1",
    "datasets>=2.14.4",
    "transformers==4.55.4",
    "einops>=0.8.0",
    "timm>=1.0.12",
    "kornia>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true

[tool.ruff]
line-length = 120
target-version = "py310"
lint.ignore = ["F401", 'E741']

[tool.ruff.lint.isort]
known-first-party = ["layerd"]

[tool.uv]
index-strategy = "unsafe-best-match"

[tool.uv.pip]
emit-index-url = true

[tool.uv.sources]
cr-renderer = { git = "https://github.com/CyberAgentAILab/cr-renderer.git", rev = "a17e1fb" }


[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.29.0",
    "matplotlib>=3.10.0",
    "pre-commit>=4.0.1",
    "pytest>=8.3.3",
    "ruff>=0.8.0",
    "mypy>=1.5.1",
    "types-PyYAML>=6.0.0",
]
